clean up project and export structure
